+++ src/Providers/CoreServiceProvider.php
@@ -18,24 +18,25 @@ use BagistoPlus\Visual\TemplateRegistrar;
 use BagistoPlus\Visual\ThemePathsResolver;
 use BagistoPlus\Visual\ThemeSettingsLoader;
 use BagistoPlus\Visual\View\Compilers\LivewireBlockCompiler;
-use Craftile\Laravel\Events\BlockSchemaRegistered;
 use Craftile\Laravel\Events\JsonViewLoaded;
 use Craftile\Laravel\Facades\Craftile;
 use Craftile\Laravel\View\BlockCompilerRegistry;
 use Illuminate\Contracts\Foundation\Application;
-use Illuminate\Contracts\Http\Kernel;
+use Illuminate\Contracts\Http\Kernel as HttpKernelContract;
 use Illuminate\Database\Eloquent\Relations\Relation;
+use Illuminate\Foundation\Http\Kernel;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\View\DynamicComponent;
 use Livewire\Component;
-use Livewire\Livewire;
 use Webkul\Attribute\Models\Attribute;
 use Webkul\Category\Models\Category;
+use Webkul\Core\Models\Channel;
 use Webkul\Product\Models\Product;
 use Webkul\Shop\Http\Middleware\Theme;
+use Webkul\Theme\Models\ThemeCustomization;
 
 class CoreServiceProvider extends ServiceProvider
 {
@@ -149,14 +150,6 @@ class CoreServiceProvider extends ServiceProvider
         if (class_exists(Component::class)) {
             $registry = app(BlockCompilerRegistry::class);
             $registry->register(new LivewireBlockCompiler);
-
-            // Register Livewire blocks as Livewire components when they're discovered
-            Event::listen(BlockSchemaRegistered::class, function (BlockSchemaRegistered $event) {
-                if ($event->schema->class && is_subclass_of($event->schema->class, Component::class)) {
-                    $componentName = 'craftile-'.$event->schema->slug;
-                    Livewire::component($componentName, $event->schema->class);
-                }
-            });
         }
     }
 
@@ -177,8 +170,9 @@ class CoreServiceProvider extends ServiceProvider
     {
         $this->app->bind(Theme::class, UseShopThemeFromRequest::class);
 
-        $this->app[Kernel::class]
-            ->prependMiddleware(DisableResponseCacheInDesignMode::class);
+        /** @var Kernel $kernel */
+        $kernel = $this->app->make(HttpKernelContract::class);
+        $kernel->prependMiddleware(DisableResponseCacheInDesignMode::class);
     }
 
     protected function bootVisualSections(): void
@@ -205,6 +199,8 @@ class CoreServiceProvider extends ServiceProvider
             'product' => Product::class,
             'category' => Category::class,
             'attribute' => Attribute::class,
+            'theme' => ThemeCustomization::class,
+            'channel' => Channel::class,
         ]);
     }